move FuncSensor class
