+++ examples/ESP8266WiFi_AT/ESP8266WiFi_AT.ino
@@ -38,7 +38,7 @@ void setup() {
 
 void loop() {
 
-  // Send using console
+  // Send AT using console
   if(Serial.available()){
     readAndSendAT();
   }
